requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6

